Split out Range, Scan and Search operations to make way for client implementation
