+++ client.go
@@ -83,133 +83,6 @@ type BatchReadResult map[DomainObject]error
 // It's a convenience function for code readability.
 func All() []string { return nil }
 
-// RangeOp is used to specify constraints to Range calls
-type RangeOp struct{}
-
-// String satisfies the Stringer interface
-func (r *RangeOp) String() string {
-	/* TODO */
-	return ""
-}
-
-// Eq is used to express an equality constraint for a range query
-func (r *RangeOp) Eq(string, interface{}) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// Gt is used to express an "greater than" constraint for a range query
-func (r *RangeOp) Gt(key string, value interface{}) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// GtOrEq is used to express an "greater than or equal" constraint for a
-// range query
-func (r *RangeOp) GtOrEq(key string, value interface{}) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// Lt is used to express a "less than" constraint for a range query
-func (r *RangeOp) Lt(key string, value interface{}) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// LtOrEq is used to express a "less than or equal" constraint for a
-// range query
-func (r *RangeOp) LtOrEq(key string, value interface{}) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// Fields list the non-key fields users want to fetch. If not set, all fields would be fetched.
-// PrimaryKey fields are always fetched.
-func (r *RangeOp) Fields([]string) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// Limit sets the number of rows returned per call. If not set, a default
-// value would be applied
-func (r *RangeOp) Limit(n int) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// Offset sets the pagination token. If not set, an empty token would be used.
-func (r *RangeOp) Offset(token string) *RangeOp {
-	/* TODO */
-	return r
-}
-
-// SearchOp represents the search query using a "searchable" field.
-type SearchOp struct {
-	/* TODO */
-}
-
-// NewSearchOp returns a new SearchOp instance
-func NewSearchOp(DomainObject) *SearchOp {
-	return &SearchOp{}
-}
-
-// String satisfies the stringer interface
-func (s *SearchOp) String() string {
-	/* TODO */
-	return ""
-}
-
-// By indicates the "searchable" field name and its value.
-func (s *SearchOp) By(fieldName string, fieldValue interface{}) *SearchOp {
-	/* TODO */
-	return s
-}
-
-// Limit sets the number of rows returned per call. Default is 128.
-func (s *SearchOp) Limit(n int) *SearchOp {
-	/* TODO */
-	return s
-}
-
-// Offset sets the pagination token. If not set, an empty token would be used.
-func (s *SearchOp) Offset(token string) *SearchOp {
-	/* TODO */
-	return s
-}
-
-// Fields list the non-key fields users want to fetch. If not set, all normalized fields
-// (supplied with “storing” annotation) would be fetched.
-// PrimaryKey fields are always fetched.
-func (s *SearchOp) Fields([]string) *SearchOp {
-	/* TODO */
-	return s
-}
-
-// ScanOp represents the scan query
-type ScanOp struct {
-	/* TODO */
-}
-
-// Limit sets the number of rows returned per call. Default is 128.
-func (s *ScanOp) Limit(n int) *ScanOp {
-	/* TODO */
-	return s
-}
-
-// Offset sets the pagination token. If not set, an empty token would be used.
-func (s *ScanOp) Offset(token string) *ScanOp {
-	/* TODO */
-	return s
-}
-
-// Fields list the non-key fields users want to fetch.
-// PrimaryKey fields are always fetched.
-func (s *ScanOp) Fields([]string) *ScanOp {
-	/* TODO */
-	return s
-}
-
 // AdminClient has methods to manage schemas and scopes
 type AdminClient interface {
 	// CheckSchema checks the compatibility of schemas
